+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 1.0.5
+* Issue [#13](https://github.com/comcduarte/ecm-frontend/issues/13) Only add vendor metadata if submitted
 * Issue [#12](https://github.com/comcduarte/ecm-frontend/issues/12) Amendments should be listed when viewing contract
 
 ## 1.0.4 - 2026-01-08